refactor: remove constants barrel file
