--- src/constants/index.ts
@@ -1,2 +1,0 @@
-export * from "./tokens"
-export * from "./aurora"
